// app/src/interfaces/hipotecaInterfaces.tsx
export interface HipotecaResponse {
  id_hipoteca: number;
  nombre_solicitante: string;
  estado_civil: string;
  fecha_nacimiento: string;
  ingreso_mensual: string;
  domicilio: string;
  empleador: string;
  antiguedad_laboral: string;
  monto_solicitado: string;
  plazo_pago: string;
  propiedad_financiar: string;
  valor_inmueble: string;
  aval_coacreditado: string;
  referencias_personales: string;
  cuenta_bancaria: string;
  firma_digital: string;
  ine_solicitante: string;
  comprobante_ingresos: string;
  avaluo_inmueble: string;
  predial_escritura: string;
  comprobante_domicilio: string;
  fotografia_propiedad: string;
}